Established working database, fixed datastoring, connections and changed fieldtypes

+++ src/main/java/com/codecool/shop/model/Product.java
@@ -7,25 +7,25 @@ import java.util.Currency;
 public class Product extends BaseModel {
 
     @Expose
-    private float defaultPrice;
+    private int defaultPrice;
     @Expose
     private Currency defaultCurrency;
     private ProductCategory productCategory;
     private Supplier supplier;
 
 
-    public Product(String name, float defaultPrice, String currencyString, String description, ProductCategory productCategory, Supplier supplier) {
+    public Product(String name, int defaultPrice, String currencyString, String description, ProductCategory productCategory, Supplier supplier) {
         super(name, description);
         this.setPrice(defaultPrice, currencyString);
         this.setSupplier(supplier);
         this.setProductCategory(productCategory);
     }
 
-    public float getDefaultPrice() {
+    public int getDefaultPrice() {
         return defaultPrice;
     }
 
-    public void setDefaultPrice(float defaultPrice) {
+    public void setDefaultPrice(int defaultPrice) {
         this.defaultPrice = defaultPrice;
     }
 
@@ -45,7 +45,7 @@ public class Product extends BaseModel {
         return id;
     }
 
-    public void setPrice(float price, String currency) {
+    public void setPrice(int price, String currency) {
         this.defaultPrice = price;
         this.defaultCurrency = Currency.getInstance(currency);
     }
